refactor: rename namespace 'Actions\*' to 'Action\*' and corresponding folder structure
