+++ src/action/ApplyAction.php
@@ -1,5 +1,5 @@
 <?php
-namespace R794021\Actions;
+namespace R794021\Action;
 
 use R794021\Tasks\Task;
 use R794021\Users\User;
