+++ src/tests/actions/ApplyActionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use R794021\Actions\ApplyAction;
+use R794021\Action\ApplyAction;
 use R794021\Tasks\Task;
 use R794021\Users\{Customer, Contractor};
