+++ src/tests/actions/CancelActionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use R794021\Actions\CancelAction;
+use R794021\Action\CancelAction;
 use R794021\Tasks\Task;
 use R794021\Users\{Contractor, Customer};
