+++ src/tests/actions/DoneActionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use R794021\Actions\DoneAction;
+use R794021\Action\DoneAction;
 use R794021\Tasks\Task;
 use R794021\Users\{Contractor, Customer};
